replace react-router's BrowserHistory with history
with respect to the example in README, #198 replaced `history.createHistory` with the `browserHistory` module from react-router... 

what is the recommended usage? should we be using react-router's internal history singleton or createHistory?

my understanding was that `redux-simple-router` is supposed to integrate with history directly (via react-router's useRouterHistory enhancer). i have updated the example to reflect this. i used a similar [implementation for another migration](https://github.com/davezuko/react-redux-starter-kit/pull/458/files).