Init currentLocation to initial location from the store
This solves the problem when the server has initialised the state and the user already has things in their history.
Without this change a superfluous action will be pushed with the same URL that the user is already looking at resulting in that we would need to use the back button twice.

Resolves #313