Makes sure the state in the store matches the state in history
This solves some problems where they would be out of sync on the first render when using server side rendering. The location would be the same for both of them but the key would be different resulting in some strange behaviour.

The application will now always dispatch a LOCATION_CHANGE on initial render. This was already done when not using server side rendering, meaning that the behaviour is similar between using server side rendering and not.